generate new stories for form inputs
