+++ libs/shared/src/lib/components/form/stories/inputs/form-datetime.stories.ts
@@ -1,0 +1,45 @@
+import {
+  FormInputStory,
+  sharedForm,
+  sharedQuestion,
+  FormInputStoryMeta,
+} from './form-inputs.stories-shared';
+
+export default {
+  ...FormInputStoryMeta,
+  title: 'Form/Inputs/DateTime',
+};
+
+/**
+ * Default inputs DateTime
+ */
+export const DateTime: FormInputStory = {
+  args: {
+    title: 'DateTime question',
+  },
+  render: (args) => {
+    return {
+      props: {
+        form: {
+          ...sharedForm,
+          structure: JSON.stringify({
+            pages: [
+              {
+                name: 'page1',
+                elements: [
+                  {
+                    type: 'text',
+                    name: 'question1',
+                    inputType: 'datetime-local',
+                    ...sharedQuestion(args),
+                  },
+                ],
+              },
+            ],
+            showQuestionNumbers: 'off',
+          }),
+        },
+      },
+    };
+  },
+};
