+++ libs/shared/src/lib/components/form/stories/inputs/form-range.stories.ts
@@ -1,0 +1,44 @@
+import {
+  FormInputStory,
+  sharedForm,
+  sharedQuestion,
+  FormInputStoryMeta,
+} from './form-inputs.stories-shared';
+
+export default {
+  ...FormInputStoryMeta,
+  title: 'Form/Inputs/Range',
+};
+
+/**
+ * Default inputs Range
+ */
+export const Range: FormInputStory = {
+  args: {
+    title: 'Range question',
+  },
+  render: (args) => {
+    return {
+      props: {
+        form: {
+          ...sharedForm,
+          structure: JSON.stringify({
+            pages: [
+              {
+                name: 'page1',
+                elements: [
+                  {
+                    type: 'text',
+                    inputType: 'range',
+                    ...sharedQuestion(args),
+                  },
+                ],
+              },
+            ],
+            showQuestionNumbers: 'off',
+          }),
+        },
+      },
+    };
+  },
+};
