+++ libs/shared/src/lib/components/form/stories/inputs/form-time.stories.ts
@@ -1,0 +1,44 @@
+import {
+  FormInputStory,
+  sharedForm,
+  sharedQuestion,
+  FormInputStoryMeta,
+} from './form-inputs.stories-shared';
+
+export default {
+  ...FormInputStoryMeta,
+  title: 'Form/Inputs/Time',
+};
+
+/**
+ * Default inputs Time
+ */
+export const Time: FormInputStory = {
+  args: {
+    title: 'Time question',
+  },
+  render: (args) => {
+    return {
+      props: {
+        form: {
+          ...sharedForm,
+          structure: JSON.stringify({
+            pages: [
+              {
+                name: 'page1',
+                elements: [
+                  {
+                    type: 'text',
+                    inputType: 'time',
+                    ...sharedQuestion(args),
+                  },
+                ],
+              },
+            ],
+            showQuestionNumbers: 'off',
+          }),
+        },
+      },
+    };
+  },
+};
