+++ libs/shared/src/lib/components/form/stories/inputs/form-url.stories.ts
@@ -1,0 +1,44 @@
+import {
+  FormInputStory,
+  sharedForm,
+  sharedQuestion,
+  FormInputStoryMeta,
+} from './form-inputs.stories-shared';
+
+export default {
+  ...FormInputStoryMeta,
+  title: 'Form/Inputs/URL',
+};
+
+/**
+ * Default inputs URL
+ */
+export const URL: FormInputStory = {
+  args: {
+    title: 'URL question',
+  },
+  render: (args) => {
+    return {
+      props: {
+        form: {
+          ...sharedForm,
+          structure: JSON.stringify({
+            pages: [
+              {
+                name: 'page1',
+                elements: [
+                  {
+                    type: 'text',
+                    inputType: 'url',
+                    ...sharedQuestion(args),
+                  },
+                ],
+              },
+            ],
+            showQuestionNumbers: 'off',
+          }),
+        },
+      },
+    };
+  },
+};
